requirements.txt
numpy==2.3.1
scikit-learn==1.7.1
yfinance==0.2.65
google-auth-oauthlib==1.2.2
google-api-python-client==2.178.0